Move AmountMicros parse check to Editor assembly.
APITests are compile-only; the regression runs on editor load instead.
