+++ RevenueCat/Editor/SubscriptionOptionPriceTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7c4e91b2d8f4e6a9b3c1f0e5d827641
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
